pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "caloron-alpibru"
version = "0.3.3"
description = "Autonomous AI sprint orchestration — agents that ship real code via Git. (Import as `caloron`.)"
readme = "README.md"
license = "EUPL-1.2"
license-files = ["LICENSE"]
requires-python = ">=3.11"
authors = [{ name = "Alfonso Sastre", email = "alfonso@alpibru.com" }]
maintainers = [{ name = "Alpibru", email = "hello@alpibru.com" }]
keywords = [
    "agent", "ai-agent", "orchestration", "sprint", "autonomous",
    "claude", "gemini", "git", "caloron", "agentspec",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "agentspec-alpibru>=0.1.0",
    "pyyaml>=6.0",
    "acli-spec>=0.4.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
caloron = "caloron.cli.main:main"

[project.urls]
Homepage = "https://www.alpibru.com"
Documentation = "https://alpibrusl.github.io/caloron-noether/"
Repository = "https://github.com/alpibrusl/caloron-noether"
Issues = "https://github.com/alpibrusl/caloron-noether/issues"

[tool.hatch.build.targets.wheel]
packages = ["caloron", "orchestrator"]

[tool.hatch.build.targets.wheel.force-include]
"scripts" = "caloron/_scripts"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["caloron", "orchestrator", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = [
    "E501",   # line length — handled by formatter
    "B008",   # function call in default arg — Typer requires this for Option/Argument
]

[tool.ruff.lint.per-file-ignores]
# Legacy orchestrator scripts predate the lint rules; don't touch them
"orchestrator/*" = ["E701", "E722", "B904", "F401"]
# Legacy stage business logic — migrated mechanically from stdin/stdout
# shape to execute(input) in v0.3.4. Don't reshape behaviour as a side
# effect of lint; the F841 unused locals were present in the originals.
"stages/dag/*" = ["F841", "B007"]
"stages/github/*" = ["F841"]
"stages/retro/*" = ["F841"]
"stages/supervisor/*" = ["F841"]
"stages/kickoff/*" = ["F841"]
# Tests can use bare except for unstable assertions
"tests/*" = ["E722"]

[tool.hatch.build.targets.sdist]
include = [
    "caloron",
    "orchestrator",
    "scripts",
    "stages",
    "compositions",
    "shell",
    "templates",
    "LICENSE",
    "README.md",
    "CHANGELOG.md",
]
exclude = [
    "**/__pycache__",
    "**/target",          # Rust build artifacts (shell/target, etc.)
    "**/*.pyc",
    ".venv",
    ".pytest_cache",
    "site",
    "demo",
    "context",
    "docs/casts",
    "dist",
]
